Add 'sonar.failOnError' property
Defaults to 'true', maintaining existing behaviour. When 'false' and the
the EmbeddedScanner throws an exception, the exception is not propagated
and is instead logged as a warning.

Setting the property to false allows the Maven build to succeed despite
problems that may be encounterted with Sonar, for example network
connectivity issues/timeouts etc...