MSONAR-205 Remove sonar.projectKey for submodules
Fix a bug where the scanner-engine would reject a multi-module project
where the `sonar.projectKey` is specified in the root or in a non-leaf
module.

When the user specifies `sonar.projectKey` as a maven property in the root (or
in any non-leaf project), the property is inherited by its sub modules.
All (sub)modules would then share the same `sonar.projectKey` clashing
with a legacy test in the scanner-engine that would ensure that each
submodule has a different key (where the key is defined as whatever
value is set for the property `sonar.projectKey` of that module).

MSONAR-205